# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Character {
  /**
     * @minLength 1
     * @maxLength 60
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 999
     */
  age: number;
  /**
     * @minLength 1
     * @maxLength 60
     */
  role: string;
  /** @maxLength 300 */
  personality: string;
  /** @maxLength 500 */
  background: string;
  /** @maxLength 200 */
  strengths: string;
  /** @maxLength 200 */
  weaknesses: string;
  /** @maxLength 300 */
  goal: string;
  /** @maxLength 300 */
  fear: string;
  /** @maxLength 200 */
  ability: string;
}

export interface Choice {
  id: string;
  text: string;
}

export interface Scene {
  id: string;
  /** @minimum 1 */
  chapter: number;
  chapterTitle: string;
  title: string;
  location: string;
  mood: string;
  narrative: string;
  /** @nullable */
  playerAction?: string | null;
  choices: Choice[];
  createdAt: string;
}

export interface WorldState {
  currentLocation: string;
  inventory: string[];
  quests: string[];
  importantEvents: string[];
  decisions: string[];
  currentObjective: string;
  discoveredCharacters?: string[];
}

export type StoryStatus = typeof StoryStatus[keyof typeof StoryStatus];


export const StoryStatus = {
  active: 'active',
  completed: 'completed',
} as const;

export interface Story {
  id: string;
  title: string;
  genre: string;
  world: string;
  tone: string;
  length: string;
  difficulty: string;
  style: string;
  character: Character;
  scenes: Scene[];
  state: WorldState;
  status: StoryStatus;
  createdAt: string;
  updatedAt: string;
}

export interface StoryInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  genre: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  world: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  tone: string;
  /**
     * @minLength 1
     * @maxLength 20
     */
  length: string;
  /**
     * @minLength 1
     * @maxLength 20
     */
  difficulty: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  style: string;
  character: Character;
}

export interface StoryUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
}

export interface StoryTurnInput {
  /**
     * @maxLength 1000
     * @nullable
     */
  action: string | null;
}

export interface StoryStats {
  total: number;
  active: number;
  completed: number;
  choices: number;
  chapters: number;
}