# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Character {
  /**
     * @minLength 1
     * @maxLength 60
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 999
     */
  age: number;
  /**
     * @minLength 1
     * @maxLength 60
     */
  role: string;
  /** @maxLength 300 */
  personality: string;
  /** @maxLength 500 */
  background: string;
  /** @maxLength 200 */
  strengths: string;
  /** @maxLength 200 */
  weaknesses: string;
  /** @maxLength 300 */
  goal: string;
  /** @maxLength 300 */
  fear: string;
  /** @maxLength 200 */
  ability: string;
}