# lib/api-zod/src/generated/types/scene.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Choice } from './choice';

export interface Scene {
  id: string;
  /** @minimum 1 */
  chapter: number;
  chapterTitle: string;
  title: string;
  location: string;
  mood: string;
  narrative: string;
  /** @nullable */
  playerAction?: string | null;
  choices: Choice[];
  createdAt: Date;
}