# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Character } from './character';
import type { Scene } from './scene';
import type { StoryStatus } from './storyStatus';
import type { WorldState } from './worldState';

export interface Story {
  id: string;
  title: string;
  genre: string;
  world: string;
  tone: string;
  length: string;
  difficulty: string;
  style: string;
  character: Character;
  scenes: Scene[];
  state: WorldState;
  status: StoryStatus;
  createdAt: Date;
  updatedAt: Date;
}