# lib/api-zod/src/generated/types/storyInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Character } from './character';

export interface StoryInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  genre: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  world: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  tone: string;
  /**
     * @minLength 1
     * @maxLength 20
     */
  length: string;
  /**
     * @minLength 1
     * @maxLength 20
     */
  difficulty: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  style: string;
  character: Character;
}