# lib/api-zod/src/generated/types/storyStats.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StoryStats {
  total: number;
  active: number;
  completed: number;
  choices: number;
  chapters: number;
}