# lib/api-zod/src/generated/types/storyTurnInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StoryTurnInput {
  /**
     * @maxLength 1000
     * @nullable
     */
  action: string | null;
}