# lib/api-zod/src/generated/types/storyUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StoryUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
}