# lib/api-zod/src/generated/types/worldState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorldState {
  currentLocation: string;
  inventory: string[];
  quests: string[];
  importantEvents: string[];
  decisions: string[];
  currentObjective: string;
  discoveredCharacters?: string[];
}